requirements.txt
anyio==3.6.2
APScheduler==3.9.1
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
click==8.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytube==12.1.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
Telethon==1.25.4
tomli==2.0.1
tornado==6.2
tzdata==2022.5
tzlocal==4.2
